+++ packages/expo-file-system/src/next/FileSystem.ts
@@ -3,6 +3,7 @@ import { ReadableStream, WritableStream } from 'web-streams-polyfill';
 import ExpoFileSystem from './ExpoFileSystem';
 import { PathUtilities } from './pathUtilities';
 import { FileSystemReadableStreamSource, FileSystemWritableSink } from './streams';
+import type { DownloadOptions } from './ExpoFileSystem.types';
 
 export class Paths extends PathUtilities {
   /**
@@ -133,8 +134,8 @@ export class File extends ExpoFileSystem.FileSystemFile {
 }
 
 // Cannot use `static` keyword in class declaration because of a runtime error.
-File.downloadFileAsync = async function downloadFileAsync(url: string, to: File | Directory) {
-  const outputPath = await ExpoFileSystem.downloadFileAsync(url, to);
+File.downloadFileAsync = async function downloadFileAsync(url: string, to: File | Directory, options?: DownloadOptions) {
+  const outputPath = await ExpoFileSystem.downloadFileAsync(url, to, options);
   return new File(outputPath);
 };